Separated classes in CombinationsFile.cs into several files.
