+++ UWRandomizerEditor/CMBdotDAT/FinalCombination.cs
@@ -1,0 +1,8 @@
+﻿namespace UWRandomizerEditor.CMBdotDAT;
+
+public class FinalCombination : ItemCombination
+{
+    public FinalCombination() : base(new FinalEntry(), new FinalEntry(), new FinalEntry())
+    {
+    }
+}
